chore: export integration tests component to be use in other packages
add integration test module as dev dependency for jerni
