+++ packages/integrations/tests/models.ts
@@ -1,0 +1,4 @@
+import BankAccountModel from "./fixtures/BankAccountModel";
+import BankAccountModel_2 from "./fixtures/BankAccountModel_2";
+
+export { BankAccountModel, BankAccountModel_2 };
