+++ packages/integrations/tests/startWorker.ts
@@ -1,10 +1,7 @@
 import { JourneyInstance } from "jerni/type";
 
-export default async function startWorker(
-  journey: JourneyInstance,
-  signal: AbortSignal,
-) {
+export default async function startWorker(journey: JourneyInstance, signal: AbortSignal) {
   for await (const events of journey.begin(signal)) {
-    console.log("events", events);
+    // console.log("events", events);
   }
 }
