+++ packages/jerni/src/events-storage/injectDatabase.ts
@@ -1,0 +1,17 @@
+import { JourneyCommittedEvent } from "jerni/type";
+import getSqliteDb from "src/events-storage/sqlite-database";
+import setup from "src/asynclocal";
+
+export interface EventDatabase {
+  getEventsFrom(lastEventId: number): Promise<JourneyCommittedEvent[]>;
+  insertEvents(events: JourneyCommittedEvent[]): Promise<void>;
+}
+
+async function getDB() {
+  return getSqliteDb();
+}
+
+const store = setup<EventDatabase>(getDB);
+
+export const injectEventDatabase = store.inject;
+export const getEventDatabase = store.get;
